Better layout for TV + created shared delegate
