+++ app/src/main/res/values/styles.xml
@@ -17,8 +17,13 @@
         <item name="android:textColorPrimary">#ff000000</item>
         <item name="android:textColorSecondary">#8a000000</item>
         <item name="colorControlNormal">#8a000000</item>
+        <item name="android:textViewStyle">@style/LeanbackTextViewStyle</item>
 
         <item name="colorPrimaryDark">@color/colorPrimaryVariant</item>
         <item name="colorAccent">@color/colorPrimary</item>
+    </style>
+
+    <style name="LeanbackTextViewStyle" parent="Widget.AppCompat.TextView">
+        <item name="android:textColor">?android:textColorSecondary</item>
     </style>
 </resources>
